Don't write version file to disk in simulation

+++ wpilib/wpilib/robotbase.py
@@ -173,12 +173,13 @@ class RobotBase:
             logger.error("If your robot class has an __init__ function, it must call super().__init__()!")
             return False
         
-        try:
-            import wpilib
-            with open('/tmp/frc_versions/FRC_Lib_Version.ini', 'w') as fp:
-                fp.write('RobotPy %s' % wpilib.__version__)
-        except:
-            logger.warning("Could not write FRC version file to disk")
+        if not hal.HALIsSimulation():
+            try:
+                import wpilib
+                with open('/tmp/frc_versions/FRC_Lib_Version.ini', 'w') as fp:
+                    fp.write('RobotPy %s' % wpilib.__version__)
+            except:
+                logger.warning("Could not write FRC version file to disk")
 
         try:
             robot.startCompetition()
